% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AR_test.R
\name{AR_test}
\alias{AR_test}
\title{Weak-IV Robust Anderson Rubin Test}
\usage{
AR_test(
  data,
  Y,
  D,
  Z,
  controls,
  FE = NULL,
  cl = NULL,
  weights = NULL,
  prec = 4,
  alpha = 0.05,
  parallel = NULL,
  cores = NULL
)
}
\arguments{
\item{data}{dataframe}

\item{Y}{outcome (string)}

\item{D}{treatment (string)}

\item{Z}{instrument (string)}

\item{controls}{control variables (character vector)}

\item{FE}{fixed effects (character vector)}

\item{cl}{Cluster name}

\item{weights}{weighting vector}

\item{prec}{precision of CI in string}

\item{alpha}{level of statitical significance}

\item{parallel}{parallel computing}

\item{cores}{number of cores}
}
\description{
Weak-IV Robust Anderson Rubin Test
}
