% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_IV.R
\name{boot_IV}
\alias{boot_IV}
\title{bootstrap IV estimator}
\usage{
boot_IV(
  data,
  Y,
  D,
  Z,
  controls = NULL,
  FE = NULL,
  cl = NULL,
  weights = NULL,
  nboots = 1000,
  parallel = TRUE,
  seed = 94305,
  cores = NULL,
  prec = 4,
  debug = FALSE,
  public = TRUE
)
}
\arguments{
\item{data}{dataframe}

\item{Y}{outcome (string)}

\item{D}{treatment (string)}

\item{Z}{instrument (string)}

\item{controls}{control variables (character vector)}

\item{show_progress_bar}{Show progress bar (default TRUE).}
}
\value{
list of results
}
\description{
bootstrap IV estimator
}
