% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltz.R
\name{dgp_zfs}
\alias{dgp_zfs}
\title{Simulates data with exclusion restriction violation of gamma with known subgroup with zero first stage}
\usage{
dgp_zfs(
  gamma = 0,
  share_zfs = 0.2,
  n = 10000,
  alpha_0 = -1,
  pi_0 = 2,
  beta_0 = 1,
  sigma_beta = 0.25
)
}
\arguments{
\item{gamma}{Magnitude of direct effect of instrument on outcome}

\item{share_zfs}{Share of sample with first stage coefficient = 0}

\item{n}{sample size}

\item{alpha_0}{mean of heterogeneous intercept in first stage}

\item{pi_0}{mean of heterogeneous slope in first stage}

\item{beta_0}{mean of heterogeneous treatment effect}

\item{sigma_beta}{SD of heterogeneous treatment effect}
}
\description{
Simulates data with exclusion restriction violation of gamma with known subgroup with zero first stage
}
