% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effF.R
\name{eff_F_stat}
\alias{eff_F_stat}
\title{Computes effective F statistic}
\usage{
eff_F_stat(ivmod, noi = T)
}
\arguments{
\item{ivmod}{fitted felm IV object with residualised Y, D, Z}

\item{noi}{noisy, defaults to TRUE}
}
\description{
Computes effective F statistic
}
