% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formula_lfe}
\alias{formula_lfe}
\title{Stitches together formula for use in felm}
\usage{
formula_lfe(Y, X, W = NULL, D = NULL, Z = NULL, C = NULL)
}
\arguments{
\item{X}{vector of controls}

\item{W}{treatment variable}

\item{D}{vector of factor variables to be partialed out}

\item{Z}{vector of instruments}

\item{C}{vector of variables cluster standard errors (multi-way permitted by LFE)}

\item{y}{The dependent variable}
}
\description{
Stitches together formula for use in felm
}
\examples{
formula_lfe(Y='mpg', X = c('hp', 'drat'), D = c('wt', 'vs'))
formula_lfe(Y='mpg', X = c('hp', 'drat'), W = 'gear', Z = c('cyl', 'carb'), D = c('wt', 'vs'), C = c('cyl', 'wt'))
}
