\docType{data}
\name{gsz}
\alias{gsz}
\title{Data from GSZ (2016)}
\format{
A data frame with 5357 rows and 11 columns.
}
\description{Data from Guiso, Sapienza, and Zingales (2016)}
\details{
The authors revisit Putnam, Leonardi, and Nanetti (1992)’s celebrated conjecture that Italian cities that achieved self-government in the Middle Ages have higher modern-day levels of social capital. More specifically, they study the effects of free city-state status on social capital as measured by the number of nonprofit organizations and organ donations per capita, and a measure of whether students cheat in mathematics. We focus on the first outcome, the number of nonprofit organizations.
}
\references{
 Guiso, Luigi, Paola Sapienza, and Luigi Zingales. 2016. "Long-Term Persistence." Journal of the European Economic Association 14 (6): 1401--36.
}
\keyword{datasets}
