\docType{data}
\name{gsz_south}
\alias{gsz_south}
\title{Data from Guiso, Sapienza, and Zingales (2016); southern Italian cities}
\format{
A data frame with 2175 rows and 11 columns.
}
\description{
The authors revisit Putnam, Leonardi, and Nanetti (1992)’s celebrated conjecture that Italian cities that achieved self-government in the Middle Ages have higher modern-day levels of social capital. More specifically, they study the effects of free city-state status on social capital as measured by the number of nonprofit organizations and organ donations per capita, and a measure of whether students cheat in mathematics. Below is a summary. We focus on the first outcome, the number of nonprofit organizations. This dataset is a subsample of southern Italian cities, which is used as a zero-first-stage sample.
}
\references{
\itemize{
  \item Guiso, Luigi, Paola Sapienza, and Luigi Zingales. 2016. "Long-Term Persistence." Journal of the European Economic Association 14 (6): 1401--36.
}
}
\keyword{datasets}
