\name{ivDiag-package}
\alias{ivDiag-package}
\docType{package}
\title{IV Estimation and Diagnostics}
\description{Conducts various estimation and diagnostic procedure for instrumental variable designs in one shot.}
\details{
  Provides estimation and diagnostic tools for instrumental variables designs, which implements the guidelines proposed in Lal et al. (2023) <arXiv:2303.11399>, including bootstrapped confidence intervals, effective F-statistic, Anderson-Rubin test, valid-t ratio test, and local-to-zero tests.

  See \code{\link{ivDiag}} for details.
}
\author{
  Apoorva Lal; Yiqing Xu
}
\references{
Lal, Apoorva, Mackenzie William Lockhart, Yiqing Xu, and Ziwen Zu. 2023. "How Much Should We Trust Instrumental Variable Estimates in Political Science? Practical Advice Based on 67 Replicated Studies." Available at: \url{https://yiqingxu.org/papers/english/2021_iv/LLXZ.pdf}
}
