% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltz.R
\name{ltz}
\alias{ltz}
\title{Estimates Local-to-Zero IV coefficients and SEs for a single instrument}
\usage{
ltz(df, W, Z, mu, sig, ivmod)
}
\arguments{
\item{df}{dataframe}

\item{W}{treatment (string)}

\item{Z}{instrument (string)}

\item{mu}{prior point estimate of direct effect of instrument on outcome}

\item{sig}{variance of above prior}

\item{ivmod}{felm iv model object}
}
\description{
Estimates Local-to-Zero IV coefficients and SEs for a single instrument
}
