% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AR_test.R
\name{AR_test}
\alias{AR_test}
\title{Weak-IV Robust Anderson Rubin Test}
\usage{
AR_test(data, Y, D, Z, X, FE, prec = 3)
}
\arguments{
\item{data}{dataframe}

\item{Y}{outcome (string)}

\item{D}{treatment (string)}

\item{Z}{instrument (string)}

\item{X}{control variables (character vector)}

\item{FE}{fixed effects (character vector)}

\item{prec}{precision of CI in string}
}
\description{
Weak-IV Robust Anderson Rubin Test
}
