% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_IV.R
\name{boot_IV}
\alias{boot_IV}
\title{bootstrap IV estimator}
\usage{
boot_IV(
  data,
  Y,
  D,
  Z,
  controls = NULL,
  FE = NULL,
  cl = NULL,
  weights = NULL,
  nboots = 1000,
  parallel = TRUE,
  seed = 94305,
  cores = NULL,
  prec = 4,
  debug = FALSE,
  sens = TRUE
)
}
\arguments{
\item{data}{dataframe}

\item{Y}{outcome (string)}

\item{D}{treatment (string)}

\item{Z}{instrument (string)}

\item{controls}{control variables (character vector)}

\item{FE}{fixed effects (character vector)}

\item{cl}{clustering column for SE (character vector)}

\item{weights}{weights name (string)}

\item{nboots}{number of bootstrap reps (1000 by default)}

\item{parallel}{boolean for parallel bootstrap (on by default)}

\item{seed}{seed}

\item{cores}{number of cores to parallelise across (defaults to all - 1)}

\item{prec}{precision of summary (4 by default)}
}
\value{
list of results
}
\description{
bootstrap IV estimator
}
