% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effF.R
\name{eff_F}
\alias{eff_F}
\title{Computes effective F statistic}
\usage{
eff_F(data, D, Y, Z, X = NULL, FE = NULL, cl = NULL, weights = NULL)
}
\arguments{
\item{data}{dataframe}

\item{D}{treatment (string)}

\item{Y}{outcome (string)}

\item{Z}{instrument (string)}

\item{FE}{fixed effects (character vector)}

\item{cl}{clustering column for SE (character vector)}

\item{weights}{weights name (string)}

\item{controls}{control variables (character vector)}
}
\description{
Computes effective F statistic
}
