% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{partialer}
\alias{partialer}
\title{partial out controls and covariates from y}
\usage{
partialer(Y, X, FE = NULL, data, weights = NULL)
}
\arguments{
\item{X}{vector of controls}

\item{data}{Dataframe}

\item{y}{The dependent variable}

\item{FEs}{fixed effects}

\item{wegihts}{name of the weighting variable}
}
\description{
partial out controls and covariates from y
}
