\name{tF}
\alias{tF}
\title{Perform the valid t-ratio procedure}
\usage{
tF(coef, se, Fstat)
}
\arguments{
\item{coef}{a 2SLS coefficient.}
\item{se}{a standard error estimate for the estimated 2SLS coefficient.}
\item{Fstat}{a first-stage partial F statistic.}
}
\description{
Perform the valid t-ratio procedure based on @lee2022valid.
}
