\name{viz_iv_dists}
\alias{viz_iv_dists}
\title{Visualise approximate sampling distributions for scalar IV coefficient}
\usage{
viz_iv_dists(out, ltz_est, xlim = NULL)
}
\arguments{
\item{out}{output from \code{ivDiag::ivDiag}.}
\item{ltz_est}{a four vector of local-to-zero estimates produced by \code{ivDiag::ltz}.}
\item{xlim}{a two-element vector specifying the range of the x-axis.}
}
\description{
Visualise approximate sampling distributions for scalar IV coefficient.
}
